BUG. `Booz_xform.calculate_modB_boozer`. Fix integer input.
    -- This bug fix handles cases where user passes integer
       input for phi or theta as in
       >>>  b.calculate_modB_boozer(4, [3, 2], [3, 4])

+++ src/booz_xform/__init__.py
@@ -15,8 +15,8 @@ class Booz_xform(Booz_xform_cpp):
           theta (array-like): The poloidal angle values.
         """
 
-        phi = np.asanyarray(phi)
-        theta = np.asanyarray(theta)
+        phi = np.asanyarray(phi, dtype=np.float)
+        theta = np.asanyarray(theta, dtype=np.float)
 
         modB = np.zeros_like(phi)
